for Thunderbird, base startup event on the last restored window's map.
Unfortunately there is not equivalent "sessionstore-windows-restored" in TB
yet.  Many events related to windows are issued before windows are fully
realized. "mail-startup-done"/"mail-tabs-session-restored" are issued by *each*
window. "console-api-log-event" is not reliable and can be issued at relatively
early stage on heavy configs. So that the best solution is probably to count
restored windows ourselves...